+++ src/Filter/BooleanFilter.php
@@ -27,31 +27,10 @@ final class BooleanFilter extends Filter
             return;
         }
 
-        $value = $data->getValue();
-
-        if (\is_array($value)) {
-            $values = [];
-            foreach ($value as $v) {
-                if (!\in_array($v, [BooleanType::TYPE_NO, BooleanType::TYPE_YES], true)) {
-                    continue;
-                }
-
-                $values[] = (BooleanType::TYPE_YES === $v) ? 1 : 0;
-            }
-
-            if (0 === \count($values)) {
-                return;
-            }
-
-            $this->applyWhere($query, $query->getQueryBuilder()->expr()->in(sprintf('%s.%s', $alias, $field), $values));
+        if (\is_array($data->getValue())) {
+            $this->filterWithMultipleValues($query, $alias, $field, $data);
         } else {
-            if (!\in_array($value, [BooleanType::TYPE_NO, BooleanType::TYPE_YES], true)) {
-                return;
-            }
-
-            $parameterName = $this->getNewParameterName($query);
-            $this->applyWhere($query, sprintf('%s.%s = :%s', $alias, $field, $parameterName));
-            $query->getQueryBuilder()->setParameter($parameterName, (BooleanType::TYPE_YES === $value) ? 1 : 0);
+            $this->filterWithSingleValue($query, $alias, $field, $data);
         }
     }
 
@@ -61,6 +40,7 @@ final class BooleanFilter extends Filter
             'field_type' => BooleanType::class,
             'operator_type' => HiddenType::class,
             'operator_options' => [],
+            'treat_null_as' => null,
         ];
     }
 
@@ -73,5 +53,52 @@ final class BooleanFilter extends Filter
             'operator_options' => $this->getOption('operator_options'),
             'label' => $this->getLabel(),
         ]];
+    }
+
+    private function filterWithMultipleValues(ProxyQueryInterface $query, string $alias, string $field, FilterData $data): void
+    {
+        $values = [];
+        foreach ($data->getValue() as $v) {
+            if (!\in_array($v, [BooleanType::TYPE_NO, BooleanType::TYPE_YES], true)) {
+                continue;
+            }
+
+            $values[] = (BooleanType::TYPE_YES === $v) ? 1 : 0;
+        }
+
+        if (0 === \count($values)) {
+            return;
+        }
+
+        $or = $query->getQueryBuilder()->expr()->orX();
+        $treatNullAs = $this->getOption('treat_null_as');
+        if (
+            false === $treatNullAs && \in_array(0, $values, true)
+            || true === $treatNullAs && \in_array(1, $values, true)
+        ) {
+            $or->add($query->getQueryBuilder()->expr()->isNull(sprintf('%s.%s', $alias, $field)));
+        }
+
+        $or->add($query->getQueryBuilder()->expr()->in(sprintf('%s.%s', $alias, $field), $values));
+
+        $this->applyWhere($query, $or);
+    }
+
+    private function filterWithSingleValue(ProxyQueryInterface $query, string $alias, string $field, FilterData $data): void
+    {
+        $or = $query->getQueryBuilder()->expr()->orX();
+        $treatNullAs = $this->getOption('treat_null_as');
+        if (
+            false === $treatNullAs && BooleanType::TYPE_NO === $data->getValue()
+            || true === $treatNullAs && BooleanType::TYPE_YES === $data->getValue()
+        ) {
+            $or->add($query->getQueryBuilder()->expr()->isNull(sprintf('%s.%s', $alias, $field)));
+        }
+
+        $parameterName = $this->getNewParameterName($query);
+        $or->add(sprintf('%s.%s = :%s', $alias, $field, $parameterName));
+        $query->getQueryBuilder()->setParameter($parameterName, (BooleanType::TYPE_YES === $data->getValue()) ? 1 : 0);
+
+        $this->applyWhere($query, $or);
     }
 }