+++ tests/Filter/BooleanFilterTest.php
@@ -59,6 +59,40 @@ final class BooleanFilterTest extends FilterTestCase
         $this->assertTrue($filter->isActive());
     }
 
+    public function testFilterNoWithTreatNullAsTrue(): void
+    {
+        $filter = new BooleanFilter();
+        $filter->initialize('field_name', [
+            'treat_null_as' => true,
+            'field_options' => ['class' => 'FooBar'],
+        ]);
+
+        $proxyQuery = new ProxyQuery($this->createQueryBuilderStub());
+
+        $filter->filter($proxyQuery, 'alias', 'field', ['type' => null, 'value' => BooleanType::TYPE_NO]);
+
+        $this->assertSameQuery(['WHERE alias.field = :field_name_0'], $proxyQuery);
+        $this->assertSameQueryParameters(['field_name_0' => 0], $proxyQuery);
+        $this->assertTrue($filter->isActive());
+    }
+
+    public function testFilterNoWithTreatNullAsFalse(): void
+    {
+        $filter = new BooleanFilter();
+        $filter->initialize('field_name', [
+            'treat_null_as' => false,
+            'field_options' => ['class' => 'FooBar'],
+        ]);
+
+        $proxyQuery = new ProxyQuery($this->createQueryBuilderStub());
+
+        $filter->filter($proxyQuery, 'alias', 'field', ['type' => null, 'value' => BooleanType::TYPE_NO]);
+
+        $this->assertSameQuery(['WHERE alias.field IS NULL OR alias.field = :field_name_0'], $proxyQuery);
+        $this->assertSameQueryParameters(['field_name_0' => 0], $proxyQuery);
+        $this->assertTrue($filter->isActive());
+    }
+
     public function testFilterYes(): void
     {
         $filter = new BooleanFilter();
@@ -73,6 +107,40 @@ final class BooleanFilterTest extends FilterTestCase
         $this->assertTrue($filter->isActive());
     }
 
+    public function testFilterYesWithTreatNullAsFalse(): void
+    {
+        $filter = new BooleanFilter();
+        $filter->initialize('field_name', [
+            'treat_null_as' => false,
+            'field_options' => ['class' => 'FooBar'],
+        ]);
+
+        $proxyQuery = new ProxyQuery($this->createQueryBuilderStub());
+
+        $filter->filter($proxyQuery, 'alias', 'field', ['type' => null, 'value' => BooleanType::TYPE_YES]);
+
+        $this->assertSameQuery(['WHERE alias.field = :field_name_0'], $proxyQuery);
+        $this->assertSameQueryParameters(['field_name_0' => 1], $proxyQuery);
+        $this->assertTrue($filter->isActive());
+    }
+
+    public function testFilterYesWithTreatNullAsTrue(): void
+    {
+        $filter = new BooleanFilter();
+        $filter->initialize('field_name', [
+            'treat_null_as' => true,
+            'field_options' => ['class' => 'FooBar'],
+        ]);
+
+        $proxyQuery = new ProxyQuery($this->createQueryBuilderStub());
+
+        $filter->filter($proxyQuery, 'alias', 'field', ['type' => null, 'value' => BooleanType::TYPE_YES]);
+
+        $this->assertSameQuery(['WHERE alias.field IS NULL OR alias.field = :field_name_0'], $proxyQuery);
+        $this->assertSameQueryParameters(['field_name_0' => 1], $proxyQuery);
+        $this->assertTrue($filter->isActive());
+    }
+
     public function testFilterArray(): void
     {
         $filter = new BooleanFilter();
@@ -81,6 +149,38 @@ final class BooleanFilterTest extends FilterTestCase
         $proxyQuery = new ProxyQuery($this->createQueryBuilderStub());
 
         $filter->filter($proxyQuery, 'alias', 'field', FilterData::fromArray(['type' => null, 'value' => [BooleanType::TYPE_NO]]));
+
+        $this->assertSameQuery(['WHERE alias.field IN ("0")'], $proxyQuery);
+        $this->assertTrue($filter->isActive());
+    }
+
+    public function testFilterArrayWithTreatNullAsFalse(): void
+    {
+        $filter = new BooleanFilter();
+        $filter->initialize('field_name', [
+            'treat_null_as' => false,
+            'field_options' => ['class' => 'FooBar'],
+        ]);
+
+        $proxyQuery = new ProxyQuery($this->createQueryBuilderStub());
+
+        $filter->filter($proxyQuery, 'alias', 'field', ['type' => null, 'value' => [BooleanType::TYPE_NO]]);
+
+        $this->assertSameQuery(['WHERE alias.field IS NULL OR alias.field IN ("0")'], $proxyQuery);
+        $this->assertTrue($filter->isActive());
+    }
+
+    public function testFilterArrayWithTreatNullAsTrue(): void
+    {
+        $filter = new BooleanFilter();
+        $filter->initialize('field_name', [
+            'treat_null_as' => true,
+            'field_options' => ['class' => 'FooBar'],
+        ]);
+
+        $proxyQuery = new ProxyQuery($this->createQueryBuilderStub());
+
+        $filter->filter($proxyQuery, 'alias', 'field', ['type' => null, 'value' => [BooleanType::TYPE_NO]]);
 
         $this->assertSameQuery(['WHERE alias.field IN ("0")'], $proxyQuery);
         $this->assertTrue($filter->isActive());